+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v0.4.0
+ - Add flags `test` command
+   - `--verbose` will print more detailed output
+   - `--no-color` will discard all colors
+   - `--concurrent [int value]` sets the maximum concurrently executed tests in `go routines`  
+ - Add default test concurrency of `runtime.NumCPU() * 5`
+
 # v0.3.0
 
  - Add `windows` release